[#1065] NextEvent·NextRemain·TodayAndNext 순수 뷰와 ViewModel 을 WidgetScenes 로 내린다
- `NextEventWidgetInlineView`·`NextEventRectangleWidgetView`·`NextRemainEventVListiew`·`TodayAndNextWidgetView` 와 그 ViewModel 계열이 `WidgetScenes` 로 갔다. 행 모델 5종(`TodayModel`·`DateModel`·`EventModel`·`MultipleEventsSummaryModel`·`UncompletedTodayTodoSummaryModel`)과 `TodayAndNextWidgetViewModelRow` 프로토콜도 함께다
- `TodayAndNextWidgetView` 의 할일 토글은 셀마다 태그 색이 달라 `(TodoEventCellViewModel, Color) -> AnyView` 로 받는다. EventList·Foremost 가 색을 안 넘기는 것과 갈리는 건 그 자리에만 셀별 색이 걸리기 때문이다
- 두 Builder 와 Provider 는 확장에 남겼다. 둘 다 확장 소유 `CalendarEvents` 로 Domain 조회 결과를 접는다
