+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/NextEventWidget/Next/NextEventWidgetTimeLineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct NextEventWidgetTimeLineProvider: TimelineProvider {
