+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/TodayAndNext/TodayAndNextWidgetViewModel+Provider.swift
@@ -12,184 +12,8 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
-
-// MARK: - TodayAndNextWidgetViewModel
-
-struct TodayAndNextWidgetViewModel {
-    var left: PageModel
-    var right: PageModel
-    var refreshAfter: TimeInterval?
-    let defaultTagColorSetting: DefaultEventTagColorSetting
-    let customTagMap: [String: any EventTag]
-    var googleCalendarColors: GoogleCalendar.Colors = .init(ownerId: "", calendars: [:], events: [:])
-    var googleCalendarTags: [String: GoogleCalendar.Tag] = [:]
-    var appleCalendarTags: [String: AppleCalendar.Tag] = [:]
-    var widgetSetting = WidgetAppearanceSettings()
-    
-    static func sample() -> TodayAndNextWidgetViewModel {
-        let now = Date()
-        let calendar = Calendar(identifier: .gregorian)
-        let today = TodayAndNextWidgetViewModel.TodayModel(
-            weekOfDay: now.text("date_form.EEEE".localized()),
-            day: calendar.component(.day, from: now)
-        )
-        let check = TodoEventCellViewModel(
-            currentTodo: TodoCalendarEvent(
-                current: TodoEvent(uuid: "check", name: "widget.events.sample::check_mailbox".localized()), isForemost: false
-            )
-        )
-        let runningEvent = ScheduleEventCellViewModel("running", name: "\("widget.events.sample::running".localized())")
-            |> \.periodText .~ .doubleText(.init(text: "08:00"), .init(text: "10:00"))
-        
-        let todayEvents = [
-            TodayAndNextWidgetViewModel.EventModel(cvm: check),
-            TodayAndNextWidgetViewModel.EventModel(cvm: runningEvent)
-        ]
-        
-        let tomorrow = TodayAndNextWidgetViewModel.DateModel(dateText: "tomorrow".localized())
-        
-        let movingEvent = ScheduleEventCellViewModel("moving", name: "widget.events.sample::moving".localized())
-            |> \.periodText .~ .singleText(.init(text: "calendar::event_time::allday".localized()))
-            |> \.eventTimeRawValue .~ .allDay(0..<10, secondsFromGMT: 0)
-            |> \.colorSource .~ EventTagId.custom("moving")
-
-        let depositEvent = TodoEventCellViewModel("deposit", name: "widget.events.sample::deposit_remain".localized())
-            |> \.periodText .~ .singleText(.init(text: "11:00"))
-            |> \.colorSource .~ EventTagId.custom("moving")
-        
-        let lunchEvent = ScheduleEventCellViewModel("lunch", name: "widget.events.sample::luch_appointment".localized())
-            |> \.periodText .~ .singleText(.init(text: "13:00"))
-        let meeting = ScheduleEventCellViewModel("meeting", name: "widget.events.sample::meeting".localized())
-            |> \.periodText .~ .doubleText(.init(text: "16:00"), .init(text: "17:00"))
-        let tomorrowEvents = [
-            TodayAndNextWidgetViewModel.EventModel(cvm: movingEvent),
-            TodayAndNextWidgetViewModel.EventModel(cvm: depositEvent),
-            TodayAndNextWidgetViewModel.EventModel(cvm: lunchEvent),
-            TodayAndNextWidgetViewModel.EventModel(cvm: meeting)
-        ]
-        
-        let defaultTagColorSetting = DefaultEventTagColorSetting(
-            holiday: "#D6236A", default: "#088CDA"
-        )
-        
-        let left = PageModel(rows: [today] + todayEvents)
-        let right = PageModel(rows: [tomorrow] + tomorrowEvents)
-        return .init(
-            left: left, right: right,
-            defaultTagColorSetting: defaultTagColorSetting,
-            customTagMap: [
-                "moving": CustomEventTag(name: "moving", colorHex: "#FFA02E")
-            ]
-        )
-    }
-}
-
-// MARK: - rows
-
-protocol TodayAndNextWidgetViewModelRow: Identifiable {
-    associatedtype ID = String
-    var rowWeight: Float { get }
-    var id: String { get }
-}
-
-extension TodayAndNextWidgetViewModel {
-    
-    struct TodayModel: TodayAndNextWidgetViewModelRow {
-        
-        let weekOfDay: String
-        let day: Int
-        var timeZonetext: String?
-        let rowWeight: Float = 2.0
-        
-        var id: String { "\(self.day)" }
-    }
-    
-    struct DateModel: TodayAndNextWidgetViewModelRow {
-        let dateText: String
-        let rowWeight: Float = 2/3
-        
-        var id: String { self.dateText }
-    }
-    
-    struct EventModel: TodayAndNextWidgetViewModelRow {
-        let cvm: any EventCellViewModel
-        var rowWeight: Float {
-            switch cvm {
-            case let todo as TodoEventCellViewModel where todo.eventTimeRawValue == nil:
-                return 2/3
-            default:
-                return self.cvm.isAlldayEvent ? 2/3 : 1
-            }
-        }
-        
-        var id: String { self.cvm.eventIdentifier }
-    }
-    
-    struct MultipleEventsSummaryModel: TodayAndNextWidgetViewModelRow {
-        
-        let tags: [EventTagId]
-        var totalCount: Int { self.tags.count }
-        let todoCount: Int
-        var nonTodoEventCount: Int { totalCount - todoCount }
-        let rowWeight: Float = 2/3
-        
-        let id: String
-        
-        init(_ rows: [EventModel]) {
-            
-            let cvms = rows.map { $0.cvm }
-            self.id = UUID().uuidString
-            
-            self.tags = cvms.map { ($0.colorSource as? EventTagId) ?? .default }
-            self.todoCount = cvms.filter { $0 is TodoEventCellViewModel }.count
-        }
-    }
-    
-    struct UncompletedTodayTodoSummaryModel: TodayAndNextWidgetViewModelRow {
-        
-        let firstTodoName: String
-        let andOtherTodosCount: Int
-        let id: String
-        
-        var rowWeight: Float { 2/3 }
-        
-        init?(_ todos: [TodoCalendarEvent]) {
-            guard !todos.isEmpty else { return nil }
-            self.id = UUID().uuidString
-            self.firstTodoName = todos[0].name
-            self.andOtherTodosCount = todos.count-1
-        }
-    }
-}
-
-// MARK: - page model
-
-extension TodayAndNextWidgetViewModel {
-    
-    struct PageModel {
-        var rows: [any TodayAndNextWidgetViewModelRow] = []
-        
-        func remainWeight(_ max: Float) -> Float {
-            return max - rows.reduce(0, { $0 + $1.rowWeight })
-        }
-    }
-}
-
-/**
- 좌측: 좌측 최상단 dayModel 채우고
- 1. 잔여 공간에 current todo 있는만큼 채우고
- 2. 그 잔여 공간에 오늘 allday 이벤트 채울수 있는만큼 채우고
- 3. 1 + 2 이벤트 총합이 잔여공간을 초과하는 경우에는 마지막행 요약형으로 변환
- 4. 이후 좌측 잔여 공간에 오늘 이벤트 채움, 공간 모자르면 오른쪽으로 넘어가고 / 남아도 별거 안함
- 
- 우측: 전체가 가용 영역
- 1. 오늘 이벤트 남은거 있으면 우선적으로 채우고
- 2. 오늘 이벤트 공간 모자르면 -> 마지막 행은 요약형으로 변환
- 3. 잔여 공간 2개 이상일떄 우선적으로 내일 이벤트 채움 -> 내일 이벤트의 경우 요약형 없음
- 4. 이후에도 공간이 2개 이상 남으면 그 다음 일자 이벤트를 채움
- 5. 공간 남으면 계속 체우고, 공간이 모질라면 마지막 날짜의 이벤트는 축약형으로 전환
- */
 
 struct TodayAndNextWidgetViewModelBuilder {
